requirements.txt
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==1.1.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0

config~=0.5.1
etcd~=2.0.2
redis~=4.0.2